Convert to monochrome honoring the Helmholtz-Kohlrausch effect.
In [1], a model for the Helmholtz-Kohlrausch [HK] effect is
proposed together with equations on how to derive the lightness L
in Luv color space from XYZ values while taking into account
the HK effect. This commit implements this conversion from color
to monochrome.

As L is the same in Luv and Lab, the result can be used without
further transformations.

The initialization issues mentioned in the last commit are also
fixed.

[1] DOI: 10.1002/(SICI)1520-6378(199712)22:6<385::AID-COL6>3.0.CO;2-R